started writing code that uses Ryacas to take derivatives and produce Jacobians.
The Jacobians work OK, but it is importatn to make all variable names compatible with yacas (no underscores).

It is also very important to convert starndard math functions to capitalized: exp() -> Exp()

When writing C code, all powers need to be written as functions: a^b -> pow(a,b).

to do this requires counting parentheses: (1+2*(1+x))^(1/(1+n)) cannot be processed using regular expressions, because the number of parentheses is unknown.

(1+x)^(a+b) would be comparatively easy, if inside the parentheses, no further parnetheses are allowed.

This counting of parentheses is actually easier in C, so we will use a C function for this (work has started).